requirements.txt
# Core dependencies
pandas>=2.0.0
python-dotenv>=1.0.0

# PDF processing
pdfplumber>=0.10.0

# Azure services
azure-ai-formrecognizer>=3.3.0
openai>=1.0.0

# Web scraping (Phase 2)
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Label generation (Phase 4)
reportlab>=4.0.0
pillow>=10.0.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0